requirements.txt
flask==3.0.3
pyjwt==2.8.0
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2
werkzeug==3.0.3
